KRACOEUS-6702: Adjusted default value for the questionnaire attribute in the protocol amendment bean to have the questionnaire module added whenever an amendment is created.
Added usage types to checks where just the usage type amendment-renewal was being used. 

Adjusted code to make Protocol Actions tab work in the same manner as the Questionnaire tab so the Protocol Actions == > Print == > Questionnaires sub-tab will render when a user enters a protocol but does not select the Questionnaire tab to answer the questions (i.e. When a user selects the Questionnaires tab, empty answerHeaders are generated and saved to the database so that subsequent methods relying on that persisted data have it available to render panels. Made Protocol Actions tab work in this same manner for questionnaires answerHeaders).