KRACOEUS-6702: Added usage types to checks where just the usage type amendment-renewal was being used.
Adjusted code to make Protocol Actions tab work in the same manner as the Questionnaire tab so the Protocol Actions == > Print == > Questionnaires sub-tab will render when a user enters a protocol but does not select the Questionnaire tab to answer the questions (i.e. When a user selects the Questionnaires tab, empty answerHeaders are generated and saved to the database so that subsequent methods relying on that persisted data have it available to render panels. Made Protocol Actions tab work in this same manner for questionnaires answerHeaders).

Added two new data elements and supporting routines for the labels and answer headers information required in the determination of the qnnr list shown on the Protocol Actions == > Print == > Questionnaires sub-tab.  These items were needed due to the fact that the business logic used to determine the questionnaire answers for the Questionnaire tab now needs to be different from the logic used to determine data shown on the Protocol Actions == > Print == > Questionnaires sub-tab.