RESKC-421:Fix modular idc currency validation to allow commas
Attempting to verify an RESSUP case, I created a new budget in existing proposal 332, added details, and then synced to the modular budget screen.
At Save, I got Errors on all 5 budget periods.
The F&A Base field in outlined in Red.
The value displayed conforms to the format. Example P1 = 192,496.73

With this error, I cannot save, cannot navigate of the screen, cannot return to proposal.
The only way I have found to save and safely exit is to delete the F&A Base value entirely for each period.

This is the first time I've synced a modular budget since the addition of commas to the numbers.