RESKC-332:when proposal is view only don't prompt user to save on close
User is confused if they enter a proposal in view mode and then select close that the modal asks if they want to save.

Steps to Reproduce
1. Login as quickstart. Create proposal.
2. Add user psmock as Viewer Document level.
3. Note proposal number and doc number. Close. Save.
4. Open incognito or private browsing window.
5. Login as psmock.
6. Use Search Proposal link (Researcher menu) to find proposal number
7. Click ‘view’ link.
8. Navigate to one of the other sections.
9. Click close on that screen. Should not get the ‘do you want to save message’.

Results: The modal asking the user if they would like to save document displays.
Expected Results: Whenever proposal is in view mode It should not generate the modal asking to save when when closing.