RESKC-477:fix to display TBN table on add budget personnel
Tried to add a TBA person to a budget; but the list does not appear and the modal window does not fully generate.
I verified that there are 7 defined TBA persons in the maintenance document.

Steps to duplicate:
Locate an in-progress proposal or create a new one.
Open or create a budget version
On the Project Personnel screen, use the +Add Personnel button to open the modal
Select TBA from the 'search for' list option

Expected result: the list of maintained To Be Named persons should present with the ability to select how many to apply.

Current result: the field updates to To Be Named, but the rest of the modal window does not generate. There is only a letter-box window containing this type field. User cannot add a TBA person to the budget.