RESKC-306:added award audit rule to check for invalid sponsor codes
1. Create an award document filling in only the required fields.
2. For the Sponsor enter an invalid code such as '010002' which after it is entered will indicate in red "not found"
3. Submit to the award e-doc to final.
RESULT: Document saves and finalizes even though the sponsor field (which is a required field) has an invalid code entered.
EXPECTED RESULT: Document should not save or finalize with a sponsor code that is not found. An error should be displayed indicating the sponsor is invalid.

The way this is described could cause some problems. If you create an award with a sponsor that later becomes inactive you wouldn't be able to maintain your document as the hard on save errors also prevent editing. This should probably be done via a validation error which would prevent submission. It should also apply to the prime sponsor field.