RESKC-519:fix STE on award navigation after submission
Logged in as Admin to Res-Demo1
Searched for existing award using account number 2533529
Opened record > clicked edit button
Maintained Description and Transaction Type on Award Tab.
Clicked on another tab to make the award change; got an STE very much like the one listed below by Victor. (input string referenced my award -For input string: "000073-00001"

This is a blocker issue in Award .

I am getting STEs when I try to navigate to the Award tab after finalizing an Award, or when I try to finalize a Time and Money document.  This is happening in Demo1 and Demo2