RESKC-365: correctly display other degree type question
Discovered in testing S2S questionnaire supporting the PHS Fellowship form:
Users answering the s2s form supporting Fellowship questionnaires presented with 2 follow up (requirement based) questions when one of five (5) specific "Other" Degree types are selected from the argument value list. The 5 questions include the full text word "other" in the question text, so question ID 100 requirement = "OTH" presents when it should not.

In the s2s form supporting Fellowship Questionnaire, there is a question branch with a series of questions that only present if the prior answer selected from an argument value list contains a certain text value. (for an “other” type of degree).
The requirement in each question in the branch is if it contains the text value ( DDOT, DOTH, MDOT, MOTH, VDOT, and OTH).

In KC, the Upper Case requirement is not being honored, so 5 of these questions are presenting both their specific match and the “OTH” question, because there is lower case text in each with the full word ‘other’ in lower case.
Requirement in KC is picking up lowercase appearance of “other” in the question text:
ID16 = MOTH
ID 17 =DOTH
ID18 = DDOT
ID19 = VDOT
ID 100 = OTH
ID 21 = MDOT

Only ID 100= OTH is working as anticipated: (OTH should only present when OTH is selected from the prior question.)