RESKC-385: Hide delete attachment button when institutional proposals are view only
A user should be required to edit the IP document before the Notes and Attachments allow 'delete' therefore delete button should be hid in view mode.

Steps to reproduce:
1. create or search for an Institutional Proposal
2. in edit mode, on the Institutional Proposal tab under Notes and Attachments, attach a file
3. fill out other required fields
4. save and blanket approve
5. re-open the IP
6. open Notes and Attachments

Results- delete button appears next to the file.
Expected Results- delete button should not appear when the IP is in view mode, user should be required to edit document first