RESKC-406:retain posted timestamp for award and IP notes and attachments
As an Award Maintainer, or
As an IP Maintainer, I need to update aspects of existing (Awards or ) Institutional Proposals.
When editing these records, my ID and Time stamp should only be added to the document, and to any new items I maintain that add them.
Currently, when I Edit an IP record or Award with existing Attachments, ALL those existing attachments update with MY user ID and the current time stamp.
Those attachments should maintain the User ID and time stamp from when they were originally added or actually last maintained. NOT with the info from the current edit.