RESKC-301:display credit splits for award unit and persons home unit on award
1. Create an Award, completing all required fields to save the document, enter the following for Lead Unit ID: IN-IN
2. Click the Contacts tab
3. Enter the following for Employee User Name: oblood
4. Click Add to add OLBOOD as a Principal Investigator
5. Click Save
6. Observe the Unit Details and Combined Credit Split subpanels
Actual Behavior: The system displays a Unit Details subpanel row for the Lead Unit defined on the Award tab (IN-IN). The system does not display any rows for the Primary Department Code defined for the Person in the Person record (BL-RCEN). The system displays a Combined Credit Split subpanel row for the one Unit defined on the Unit Details subpanel (which is the Lead Unit entered on the Award tab).
Expected Behavior: The system should display a Unit Details subpanel row for the Lead Unit defined on the Award tab and a row for the Primary Department Code defined for the Person in the Person record. The system should display a Combined Credit Split subpanel row for each Unit defined on the Unit Details subpanel, one for the Lead Unit entered on the Award tab and one for the Primary Department Code defined for the person in the Person record.