RESKC-542:added dialog to edit notification message and subject when sending certification notification for single users
As a system Admin I set the 'Prompt User" option to "on" for the Maintenance Documents> Notification >Proposal Person Certification Request Notification)
Currently: the Notification is not generating in proposals for the user to customize the message since the addition of RESKC-504 contribution.

Steps:
Create a proposal.
Add several Persons to Key Personnel (PI, several Co-I's).
Click the "Notify" option on a person row.
*Result*: Notification is immediately sent.
*Desired Result*: If Prompt User flag is set to On for this notification, a notification window should present to the user allowing the standard notification options.