RESKC-682:fixes issues with differences in total direct cost between award summary and award parameters
Edited an award node to add funds; submitted.
Edited the T&M doc to fund the node; submitted.
On Budget Versions tab: added a new budget > opened.
Linked P1 of the proposal budget (details did not come forward).
Manually added personnel and non-personnel to get close to the limit.
On the Summary screen, edited the F&A OH override field to meet the obligated total for this budget & saved.
On Award Budget Actions > tried to submit.
Results: got an error that the direct cost was exceeded.
On Parameter screen> the direct cost amount did not reflect the details of the budget.

Desired result: the values on the parameter screen should update to reflect the amounts in the detailed budget.