RESOPS-317:Remove duplicate T&M Docs when retreiving based on aai
When doing query by criteria searches referencing a collection, it appears to return duplicate T&M results. This removes the duplication before returning from the dao.
Additionally when hitting reload on the new T&M History tab the necessary data to render the page was not being populated on the BO.