RESKC-847:Award budget line items imported with wrong dates
PB expenses where end dates are not period end dates shows wrong amounts pulled into AB.
When personnel is added to a proposal budget in a way that the end date for personnel expense is not the same as the period end date, it is not converting properly when adding the award budget. The award budget shows those end dates as the period end date and causes an STE when user tries to correct the end dates.