RESKC-961:Viewing PD causes whitescreen when no keywords defined
When viewing a PD in readonly mode(from the PD lookup or after enroute/final) and keywords are enabled and there are no keywords defined in the system a stacktrace is generated in the logs and only a whitescreen is displayed to the user. This fixes that and the read-only display of the keywords which was previously broken as well. The following is a snippet from the stack trace

For "${...}" content: Expected a string or something automatically convertible to string (number, date or boolean), but this evaluated to a sequence (wrapper: f.t.SimpleSequence):
==> (spring.status.value?default(""))  [in template "krad/WEB-INF/ftl/components/field/dataInputField.ftl" at line 64, column 27]
...
freemarker.core.NonStringException: [... Exception message was already printed; see it above ...]