RESOPS-557:Submitting T&M document without a transaction zeros out obligated and anticipated totals in some cases
When the MAKE_AWD_CUM_ANTICIPATED_OBL_EDITABLE parameter is set to Y or the amounts on the T&M hierarchy panel are un-editable for other reasons, saving or submitting the T&M document will cause a new transaction to be errornously created that zeros out the totals. This prevents this by not creating single-node transactions based on changes when those fields are disabled.