RESKC-868: Making sure that Unrecovered F and A are merged when the fiscal year, account, rate, oncampus flag are the same across budgets.  This fixes hierarchy sync and avoids the below scenario where Institutional Proposals are generated during PD routing.
Hierarchy proposal parent was able to route all the way to the approving liaison(final stop). When the liaison clicked Approve, the following presented STE error "A row duplicates another. Enter different Unrecovered F&A information."

Steps to reproduce:
Create a proposal with basic information
Key person: PI
Create a Budget: add non-personnel line item, in Details, uncheck the Apply MTDC rate flag.
Autogenerate periods.
On Commitments > UR . enter source code as "123" in each row for on campus rate, and distribute the $.
Complete the budget
Return to the proposal >
Toolbar > Hierarchy > create a parent with this proposal as sub-budget. Make note of the parent proposal number.
Toolbar > Copy > copy this proposal including the budget.
In the new proposal > toolbar > budget versions > mark the budget as for submission and complete. (it is identical to first proposal, including the UR source code).
Toolbar > Hierarchy > link this as sub-budget, child to the parent created initially.
Save & close the proposal
Open the Parent proposal, turn on Validations and make the necessary updates to finalize the proposal.
Submit this parent.
Open the route log on the Summary/Submit screen and note the next approver . close
Log in as that next approver and approve.
Repeat this step until you get to the final approver (shulte in res-demo 1)
Approve as Schulte:
Result: The STE occurs here - A row duplicates another. Enter different Unrecovered F&A information.
As well as the full STE Text (see comments).
Desired result:
There should be no restrictions on the use of duplicate sources for any of the Commitments - U/R orCost Sharing (not tested)